make getDefaultSettings async
